backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
alembic>=1.10.0
httpx>=0.24.0
psycopg2-binary>=2.9.5  # For PostgreSQL
python-multipart>=0.0.6  # For form data parsing
gunicorn>=20.1.0  # For production deployment
pyjwt>=2.6.0  # For JWT token handling
pydantic-settings>=2.0.0  # For settings management
